Project Finished; Ready for Review
After quite some assistance, I was able to finish the complex for loop within the lesson, but was given a ton of HTML code inputed within the "cart" divs as a reward. With all the declared formulas and calculations working to suit to create the envisioned customer page, it's safe to say that this project is completed! (and yes, that also includes proper documentation)